Overwrite old with new and remove WIP
